Added unit tests for getAnnotations method
