+++ server/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@graphql-tools/mock": "^8.1.3",
     "@graphql-tools/schema": "^7.1.5",
+    "@types/mongoose": "^5.11.97",
     "apollo-server-express": "3.x",
     "axios": ">=0.21.2",
     "dotenv": "^10.0.0",
@@ -51,6 +52,7 @@
     "graphql-middleware": "^6.0.10",
     "graphql-subscriptions": "^1.2.1",
     "keycloak-connect": "14.0.0",
+    "mongodb-memory-server": "^7.4.3",
     "mongoose": "^6.0.8",
     "subscriptions-transport-ws": "^0.9.19",
     "winston": "^3.3.3",
